+++ UCR/Views/MainWindow.xaml.cs
@@ -249,5 +249,10 @@ namespace HidWizards.UCR.Views
             Action showAction = () => win.Show();
             Dispatcher.BeginInvoke(showAction);
         }
+
+        private void Close_OnClick(object sender, RoutedEventArgs e)
+        {
+            Close();
+        }
     }
 }