+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Button to Axis` parameters changed to two axis values and option for initialization
 - Plugin updates are now of type short instead of long. Some operations are performed using int, to avoid wrap-around or crashes.
 - Subscription and Bind Mode callbacks are now executed as Tasks and are an Action<short> rather than dynamic
+- Default blocking to true while UCR GUI does  not support selecting block
 
 ## [0.6.0] - 2018-12-03