+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.7.0] - 2019-01-03
+
 ### Changed
 
 - `Button to Axis` parameters changed to two axis values and option for initialization